Create app instance
